+++ modules/canvaskit/helper.js
@@ -153,6 +153,23 @@ function isCanvasKitColor(ob) {
 function toUint32Color(c) {
   return ((clamp(c[3]*255) << 24) | (clamp(c[0]*255) << 16) | (clamp(c[1]*255) << 8) | (clamp(c[2]*255) << 0)) >>> 0;
 }
+// Accepts various colors representations and converts them to an array of int colors.
+// Does not handle builders.
+function assureIntColors(arr) {
+  if (arr instanceof Float32Array) {
+    var count = Math.floor(arr.length / 4);
+    var result = new Uint32Array(count);
+    for (var i = 0; i < count; i ++) {
+      result[i] = toUint32Color(arr.slice(i*4, (i+1)*4));
+    }
+    return result;
+  } else if (arr instanceof Uint32Array) {
+    return arr;
+  } else if (arr instanceof Array && arr[0] instanceof Float32Array) {
+    return arr.map(toUint32Color);
+  }
+  
+}
 function uIntColorToCanvasKitColor(c) {
     return CanvasKit.Color(
      (c >> 16) & 0xFF,
@@ -257,6 +274,36 @@ function copy2dArray(arr, dest, ptr) {
     }
   }
   return ptr;
+}
+
+// Copies an array of colors to wasm, returning an object with the pointer
+// and info necessary to use the copied colors.
+// Accepts either a flat Float32Array, flat Uint32Array or Array of Float32Arrays.
+// If color is an object that was allocated with CanvasKit.Malloc, it's pointer is
+// returned and no extra copy is performed.
+// Array of Float32Arrays is deprecated and planned to be removed, prefer flat
+// Float32Array
+// TODO(nifong): have this accept color builders.
+function copyFlexibleColorArray(colors) {
+  var result = {
+    colorPtr: nullptr,
+    count: colors.length,
+    colorType: CanvasKit.ColorType.RGBA_F32,
+  }
+  if (colors instanceof Float32Array) {
+    result.colorPtr = copy1dArray(colors, "HEAPF32");
+    result.count = colors.length / 4;
+
+  } else if (colors instanceof Uint32Array) {
+    result.colorPtr = copy1dArray(colors, "HEAPU32");
+    result.colorType = CanvasKit.ColorType.RGBA_8888;
+
+  } else if (colors instanceof Array && colors[0] instanceof Float32Array) {
+    result.colorPtr = copy2dArray(colors, "HEAPF32");
+  } else {
+    throw('Invalid argument to copyFlexibleColorArray, Not a color array '+typeof(colors));
+  }
+  return result;
 }
 
 var defaultPerspective = Float32Array.of(0, 0, 1);