+++ modules/canvaskit/CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+ - In all places where color arrays are accepted (gradient makers, drawAtlas, and MakeSkVertices),
+   You can now provide either flat Float32Arrays of float colors, Uint32Arrays of int colors, or 
+   2d Arrays of Float32Array(4) colors. The one thing you should not pass is an Array of numbers,
+   since canvaskit wouldn't be able to tell whether they're ints or floats without checking them all.
+   The fastest choice for gradients is the flat Float32Array, the fastest choice for drawAtlas and
+   MakeSkVertices is the flat Uint32Array.
+ - Color arrays may also be objects created with CanvasKit.Malloc
+
 ## [0.16.2] - 2020-06-05
 
 ### Fixed